requirements.txt
colorama
python-dotenv
flask==3.0.3
gunicorn==22.0.0
